+++ tests/Integration/Execution/MutationExecutor/BelongsToTest.php
@@ -282,7 +282,7 @@ class BelongsToTest extends DBTestCase
             'name' => 'foo',
         ]);
 
-        $this->graphQL('
+        $this->graphQL(/* @lang GraphQL */ '
         mutation {
             createTask(input: {
                 name: "foo"
@@ -321,7 +321,7 @@ class BelongsToTest extends DBTestCase
             'name' => 'foo',
         ]);
 
-        $this->graphQL('
+        $this->graphQL(/* @lang GraphQL */ '
         mutation {
             upsertTask(input: {
                 id: 1
@@ -451,7 +451,11 @@ class BelongsToTest extends DBTestCase
 
     public function testCanCreateUsingUpsertAndDisconnectBelongsTo(): void
     {
-        factory(User::class)->create();
+        /** @var \Tests\Utils\Models\User $user */
+        $user = factory(User::class)->create();
+        $task = $user->tasks()->save(
+            factory(Task::class)->make()
+        );
 
         $this->graphQL(/* @lang GraphQL */ '
         mutation {
@@ -480,12 +484,12 @@ class BelongsToTest extends DBTestCase
         ]);
 
         $this->assertTrue(
-            User::find(1)->exists,
+            User::find($user->id)->exists,
             'Must not delete the second model.'
         );
 
         $this->assertNull(
-            Task::find(1)->user,
+            $task->refresh()->user_id,
             'Must disconnect the parent relationship.'
         );
     }
@@ -532,7 +536,6 @@ class BelongsToTest extends DBTestCase
             'This model should be deleted.'
         );
 
-        $this->markTestSkipped('There is a lingering bug in here that the previous assertion did not catch.');
         $this->assertNull(
             $task->refresh()->user_id,
             'Must disconnect the parent relationship.'