+++ tests/Unit/Execution/Arguments/ArgumentSetTest.php
@@ -123,6 +123,24 @@ class ArgumentSetTest extends TestCase
         );
     }
 
+    public function testAddValueAtRootLevel(): void
+    {
+        $set = new ArgumentSet();
+        $set->addValue('foo', 42);
+
+        $this->assertSame(42, $set->arguments['foo']->value);
+    }
+
+    public function testAddValueDeep(): void
+    {
+        $set = new ArgumentSet();
+        $set->addValue('foo.bar', 42);
+
+        $foo = $set->arguments['foo']->value;
+
+        $this->assertSame(42, $foo->arguments['bar']->value);
+    }
+
     public function testRenameInput(): void
     {
         $firstName = new Argument();